+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.2] - Unreleased
+* Added nested resource denormalization
 * Fixed lowest symfony/property-access version to 4.4
 * Replaced Travis CI by Github actions
 * Removed PHPCPD QA checks